pyproject.toml
[project]
name = "deepfabric"
version = "3.0.0"
description = "Large Scale Topic based Synthetic Data Generation for LLM Fine-Tuning & Training"
authors = [{name = "Luke Hinds", email = "luke@rdrocket.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub==0.36.0",
    "datasets>=3.0,<5.0",
    "kagglehub>=0.3.0",
    "outlines==1.2.8",
    "openai>=1.107.2",
    "anthropic>=0.67.0",
    "PyYAML>=6.0.1",
    "click>=8.1.7",
    "pydantic>=2.0.0",
    "mermaid-py>=0.8.0",
    "rich>=13.0.0",
    "google-genai>=1.36.0",
    "posthog>=3.0.0",
    "packaging>=25.0",
    "transformers>=4.57.1",
    "torch>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "requests-mock>=1.11.0",
    "ruff>=0.1.0",
    "bandit>=1.7.10",
    "mermaid-py>=0.2.0",
]
docs = [
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.30.0",
]

[project.scripts]
deepfabric = "deepfabric.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deepfabric"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore::ImportWarning",
    "ignore::pydantic.PydanticDeprecatedSince20",
]
markers = [
    "asyncio: mark tests that require an asyncio event loop",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "YTT", # flake8-2020
    "S",   # flake8-bandit
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "ERA", # eradicate
    "PLC", # pylint
    "PLE", # pylint-errors
    "PLR", # pylint-refactor
    "PLW", # pylint-warnings
    "TRY", # tryceratops
]

ignore = [
    "E501",      # Line too long (handled by formatter)
    "ISC001",    # 
    "B008",      # Do not perform function call in argument defaults
    "C901",      # Function is too complex
    "T201",      # Print found
    "S101",      # Use of assert detected
    "PLR0913",   # Too many arguments to function call
    "PLR0915",   # Too many statements
    "ERA001",    # Commented out code
    "TRY003",    # Avoid specifying long messages outside the exception class
    "PLR0912",   # Too many branches
]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR0915"]
"conftest.py" = ["S101"]

[tool.ruff.lint.isort]
known-first-party = ["deepfabric"]
lines-between-types = 1

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"
